chore: 폴더 구조 및 파일 위치 변경
