+++ ContactManager/ContactManager/Core/Sources/Enum/Constants.swift
