+++ ContactManager/ContactManager/Core/Sources/Enum/Errors.swift
